use class
